docs/archive/cleanup/20260314_235801/filesystem_snapshot/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terminal-zero-governance-control-plane"
version = "0.1.0"
description = "Script-driven AI engineering governance control plane."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "Terminal Zero Maintainers" }]
dependencies = [
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<10",
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
testpaths = ["tests"]
